# [FIX] point_of_sale: restrict expensive relations from loading